gh-93649: Split gc- and allocation tests from _testcapimodule.c
